Trivial change to trigger new build with no git merge conflict.
The CI job keeps failing because of the following error:

error: Your local changes to the following files would be overwritten
by merge:
	_docs/ubuntu16-ami-list.md
Please commit your changes or stash them before you merge.

The root issue is that I’m running a “rebuild” which originally
generated two follow-on commits. When it goes to “git pull” these
commits, there’s a merge conflict. This is fixed by creating a new
build.