Trivial commit to trigger Packer build.
Everything was workign fine before, except that I negelected to update
the AWS creds in the CircleCI build. As a result, the “find AMI”
example failed, and the whole build failed.

The solution is to (1) update the CircleCI AWS creds, (2) create a
branch “publish-amis”, and (3) submit this commit in said branch to
trigger a build.